update file explorer

+++ rust/src/websocket_client/fs_api.rs
@@ -1,4 +1,5 @@
 use serde_json::json;
+use std::path::Path;
 use tokio::{fs::OpenOptions, io::AsyncWriteExt};
 
 use super::components::{path_like_to_path, ArgumentsError};
@@ -142,7 +143,23 @@ pub async fn fs_cp(
                     path_like_to_path(dest_path_like),
                 ) {
                     (Some(src), Some(dest)) => {
-                        tokio::fs::copy(src, dest).await?;
+                        let metadata = tokio::fs::metadata(src.clone()).await?;
+                        if metadata.is_file() {
+                            tokio::fs::copy(src, dest).await?;
+                        } else if metadata.is_dir() {
+                            copy_dir_all(src, dest)?;
+                        } else if metadata.is_symlink() {
+                            let real_src = tokio::fs::read_link(src.clone()).await?;
+                            if real_src.is_file() {
+                                tokio::fs::copy(src, dest).await?;
+                            } else if real_src.is_dir() {
+                                copy_dir_all(src, dest)?;
+                            } else {
+                                return Err(Box::new(ArgumentsError(Some(format!(
+                                    "Not supported file type"
+                                )))));
+                            }
+                        }
                         return Ok(serde_json::Value::Null);
                     }
                     _ => {}
@@ -151,6 +168,20 @@ pub async fn fs_cp(
         }
     }
     return Err(Box::new(ArgumentsError(None)));
+}
+
+fn copy_dir_all(src: impl AsRef<Path>, dst: impl AsRef<Path>) -> std::io::Result<()> {
+    std::fs::create_dir(&dst)?;
+    for entry in std::fs::read_dir(src)? {
+        let entry = entry?;
+        let ty = entry.file_type()?;
+        if ty.is_dir() {
+            copy_dir_all(entry.path(), dst.as_ref().join(entry.file_name()))?;
+        } else if ty.is_file() {
+            std::fs::copy(entry.path(), dst.as_ref().join(entry.file_name()))?;
+        }
+    }
+    Ok(())
 }
 
 pub async fn fs_write_file(
